pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "\"Page Analyzer\" is a website that analyzes specified pages for their SEO suitability, similar to PageSpeed Insights."
authors = ["Nikolai <nicolas.gordienko@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.1"
flake8 = "^6.0.0"
validators = "^0.20.0"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
flask-wtf = "^1.1.1"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"

[tool.black]
line-length = 88
exclude = '/(\.git|__pycache__|\.pytest_cache|\.venv|venv)/'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
